+++ myBlog/src/fr/epsi/jeeProject/servlets/UtilisateurServlet.java
@@ -8,6 +8,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.sql.SQLException;
 
@@ -18,6 +19,18 @@ public class UtilisateurServlet extends HttpServlet {
     }
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        HttpSession session = request.getSession();
+
+        if(session.getAttribute("user_email") == null) {
+            this.getServletContext().getRequestDispatcher("/index.jsp").forward(request, response);
+            return;
+        }
+
+        if (!(boolean)session.getAttribute("user_isAdmin")) {
+            this.getServletContext().getRequestDispatcher("/NotAccess.jsp").forward(request, response);
+            return;
+        }
+
         if (request.getParameter("delete") != null) {
             try {
                 new UtilisateurDao().deleteUtilisateur(new Utilisateur(request.getParameter("delete")));
